--- src/mls_crypto/service/mod.rs
@@ -1,8 +1,0 @@
-//! [`MlsService`] trait ([`api`]) and [`OpenMlsService`] OpenMLS backend.
-
-mod api;
-mod backend;
-mod openmls;
-
-pub use api::{CIPHERSUITE, DEFAULT_COMMIT_BATCH_MAX, MlsService};
-pub use openmls::OpenMlsService;
